requirements.txt
streamlit~=1.23.1
PyPDF2~=3.0.1
langchain~=0.0.198
python-dotenv~=1.0.0
pinecone-client~=2.2.2
openai
tiktoken